requirements.txt
flask
flask-cors
openai
numpy==1.24.3
spacy==3.7.4
presidio-analyzer==2.2.34
presidio-anonymizer==2.2.34
thinc==8.2.3
python-dotenv==1.0.0
gunicorn==21.2.0
